move sim streaming from canvas
