+++ src/components/FloatingPlayPause.tsx
@@ -1,16 +1,61 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { CircleLoader } from 'react-spinners';
 
 import { PlayIcon, StopIcon } from '@heroicons/react/24/outline';
 
 import { getSimulationOutput, uploadNetwork } from '~/api/network';
+import { extractCarsFromSumoMessage } from '~/helpers/sumo';
+import { useSimulation } from '~/hooks/useSimulation';
+import {
+  BASE_SIMULATION_DATA_TOPIC,
+  BASE_SIMULATION_DESTINATION_PATH,
+  BASE_SIMULATION_ERROR_TOPIC,
+  SIMULATION_SOCKET_URL,
+} from '~/simulation-urls';
+import { useCarsStore } from '~/zustand/useCarStore';
 import { useNetworkStore } from '~/zustand/useNetworkStore';
 import { usePlaying } from '~/zustand/usePlaying';
 
 export const FloatingPlayPause = () => {
   const [loading, setLoading] = useState(false);
   const network = useNetworkStore();
+  const carStore = useCarsStore();
   const player = usePlaying();
+  const { subscribe, publish, isConnected } = useSimulation({
+    brokerURL: SIMULATION_SOCKET_URL,
+  });
+
+  // streaming of simulation data
+  useEffect(() => {
+    const SIMULATION_DATA_TOPIC = `${BASE_SIMULATION_DATA_TOPIC}/${player.simulationId}`;
+    const SIMULATION_ERROR_TOPIC = BASE_SIMULATION_ERROR_TOPIC.replace(
+      '_',
+      player.simulationId ?? '',
+    );
+    const SIMULATION_DESTINATION_PATH = `${BASE_SIMULATION_DESTINATION_PATH}/${player.simulationId}`;
+
+    if (player.isPlaying && isConnected) {
+      console.warn('Subscribing to simulation data');
+
+      subscribe(SIMULATION_DATA_TOPIC, message => {
+        const data = extractCarsFromSumoMessage(message);
+
+        if (data) {
+          carStore.setCars(data);
+        }
+      });
+      subscribe(SIMULATION_ERROR_TOPIC, message => {
+        console.error(message);
+      });
+
+      publish(SIMULATION_DESTINATION_PATH, { status: 'START' });
+    } else if (!player.isPlaying && isConnected) {
+      console.warn('Unsubscribing from simulation data');
+      publish(SIMULATION_DESTINATION_PATH, { status: 'STOP' });
+    } else if (!player.isPlaying && player.simulationId) {
+      player.changeSimulationId(null);
+    }
+  }, [player.isPlaying]);
 
   const handleUpload = async () => {
     try {
@@ -47,15 +92,22 @@ export const FloatingPlayPause = () => {
   };
 
   const handleOutput = async () => {
-    player.pause();
+    try {
+      setLoading(true);
+      player.pause();
 
-    if (!player.simulationId) {
-      return;
+      if (!player.simulationId) {
+        return;
+      }
+
+      const simOutput = await getSimulationOutput(player.simulationId);
+
+      console.log({ simOutput });
+    } catch (error: unknown) {
+      console.error(error);
+    } finally {
+      setLoading(false);
     }
-
-    const simOutput = await getSimulationOutput(player.simulationId);
-
-    console.log({ simOutput });
   };
 
   return (
