+++ src/simulation-urls.ts
@@ -1,6 +1,6 @@
-export const DOMAIN_NAME = 'localhost:8080';
-export const BASE_URL = `http://${DOMAIN_NAME}`;
-export const SIMULATION_SOCKET_URL = `ws://${DOMAIN_NAME}/simulation-socket`;
+export const DOMAIN_NAME = 'sumo-server.urbanflo.app';
+export const BASE_URL = `https://${DOMAIN_NAME}`;
+export const SIMULATION_SOCKET_URL = `wss://${DOMAIN_NAME}/simulation-socket`;
 export const BASE_SIMULATION_DATA_TOPIC = '/topic/simulation';
 export const BASE_SIMULATION_ERROR_TOPIC = '/topic/simulation/_/error';
 export const BASE_SIMULATION_DESTINATION_PATH = '/app/simulation';
